requirements.txt
flask
line-bot-sdk
requests
ray
deepface
Pillow
ultralytics
opencv-python-headless
pyodbc
pyngrok
imgurpython
